sunxi:nand: add support for samsung K9GBG08U0M
This data was taken from a newer version of libnand, which is only
available as a binary blob. The flags gave me a bit of trouble, but
i verified that all of the ones in the current libnand are still
being used the same way in the new libnand.

Works as well as could be expected :)

Signed-off-by: Luc Verhaegen <libv@skynet.be>