sunxi-hdmi: Add 2 pixelclock overrides for 1280x800 hdmi projectors
The sunxi video PLL-s can unfortunately not make any random clock encountered
in EDID info. So hdmi_edid.c has a table to override the pixelclocks in some
EDID provided modes to a value which the PLL-s can actually make close to the
real value.

This commit adds the preferred EDID mode from 2 hdmi projectors to this
table, so that sunxi-hdmi will work properly with these projectors.

Reported-by: Jonathan Liu <net147@gmail.com>
Tested-by: Jonathan Liu <net147@gmail.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>